refactor: line 객체 메소드를 manager 로 위임
