+++ src/main/java/nextstep/app/LadderGameApp.java
@@ -14,7 +14,7 @@ public class LadderGameApp {
         Participants participants = new Participants(getPersons());
         LadderResults ladderResults = new LadderResults(getLadderResults());
         LadderHeight ladderHeight = new LadderHeight(getLadderHeight());
-        LadderGameManager ladderGameManager = new LadderGameManager(getLines(participants, ladderHeight));
+        LadderGameManager ladderGameManager = new LadderGameManager(participants, getLines(participants, ladderHeight));
         OutputView.printLadder(participants, ladderGameManager, ladderResults);
         String personForResult = InputView.receivePersonForResult();
         if (personForResult.equals("all")) {
@@ -42,17 +42,17 @@ public class LadderGameApp {
         return LineFactory.createLines(participants, ladderHeight);
     }
 
-    private static LadderResult getLadderResult(Participants participants, LadderResults ladderResults, LadderGameManager ladderGameManager, String personForResult) {
-        int startTrackNumber = participants.getTrackNumberByPersonName(personForResult);
-        int finishTrackNumber = ladderGameManager.start(startTrackNumber);
-        return ladderResults.getLadderResult(finishTrackNumber);
-    }
-
     private static LadderResults getLadderTotalResults(Participants participants, LadderResults ladderResults, LadderGameManager ladderGameManager) {
         return new LadderResults(participants.getPersons()
                 .stream()
                 .map(person -> getLadderResult(participants, ladderResults, ladderGameManager, person.getName()))
                 .collect(Collectors.toList()));
+    }
+
+    private static LadderResult getLadderResult(Participants participants, LadderResults ladderResults, LadderGameManager ladderGameManager, String personForResult) {
+        int startTrackNumber = participants.getTrackNumberByPersonName(personForResult);
+        int finishTrackNumber = ladderGameManager.start(startTrackNumber);
+        return ladderResults.getLadderResult(finishTrackNumber);
     }
 
 }
